+++ steel/src/lib.rs
@@ -32,7 +32,7 @@ pub mod port;
 pub mod structs;
 
 pub mod contracts;
-pub mod new_parser;
+pub mod parser;
 // pub mod vm;
 
 // #[cfg(test)]
