+++ steel/src/parser/visitors.rs
@@ -1,4 +1,4 @@
-use crate::new_parser::ast::*;
+use crate::parser::ast::*;
 use crate::rvals::Result;
 
 pub trait VisitorMut {
